chore: core-client not depending on kms/testing
